Moved activities by category api from the activity repository pattern to the category repository pattern
