+++ OutTheGC/Endpoints/ActivityEndpoint.cs
@@ -210,42 +210,6 @@ public static class ActivityEndpoint
         .WithOpenApi()
         .Produces<Activity>(StatusCodes.Status200OK)
         .Produces(StatusCodes.Status204NoContent);
-
-        group.MapGet("/category/{categoryId}/activities", async (IActivityService activityService, Guid categoryId, Guid tripId) =>
-        {
-            var activitiesByCategory = await activityService.GetActivitiesByCategoriesAsync(categoryId, tripId);
-
-            if (!activitiesByCategory.Any())
-            {
-                return Results.NotFound(new
-                {
-                    message = "There are no activities in this category."
-                });
-            }
-
-            return Results.Ok(activitiesByCategory.Select(ac => new
-            {
-                ac.Id,
-                ac.Title,
-                ac.Notes,
-                ac.Location,
-                ac.Date,
-                ac.Duration,
-                ac.Cost,
-                ac.Category,
-                ac.WebsiteUrl,
-                User = new
-                {
-                    ac.User.FullName
-                },
-                ac.CreatedAt,
-                ac.UpdatedAt,
-                VoteCount = ac.Votes.Count()
-            }));
-        })
-        .WithOpenApi()
-        .Produces<List<Activity>>(StatusCodes.Status200OK)
-        .Produces(StatusCodes.Status404NotFound);
     }
 }
