+++ OutTheGC/Endpoints/CategoryEndpoint.cs
@@ -53,6 +53,42 @@ public static class CategoryEndpoint
         .WithOpenApi()
         .Produces<Category>(StatusCodes.Status200OK)
         .Produces(StatusCodes.Status204NoContent);
+
+        group.MapGet("/category/{categoryId}/activities", async (ICategoryService categoryService, Guid categoryId, Guid tripId) =>
+        {
+            var activitiesByCategory = await categoryService.GetActivitiesByCategoriesAsync(categoryId, tripId);
+
+            if (!activitiesByCategory.Any())
+            {
+                return Results.NotFound(new
+                {
+                    message = "There are no activities in this category."
+                });
+            }
+
+            return Results.Ok(activitiesByCategory.Select(ac => new
+            {
+                ac.Id,
+                ac.Title,
+                ac.Notes,
+                ac.Location,
+                ac.Date,
+                ac.Duration,
+                ac.Cost,
+                ac.Category,
+                ac.WebsiteUrl,
+                User = new
+                {
+                    ac.User.FullName
+                },
+                ac.CreatedAt,
+                ac.UpdatedAt,
+                VoteCount = ac.Votes.Count()
+            }));
+        })
+        .WithOpenApi()
+        .Produces<List<Activity>>(StatusCodes.Status200OK)
+        .Produces(StatusCodes.Status404NotFound);
     }
 }
