+++ OutTheGC/Interfaces/IActivityRepository.cs
@@ -21,6 +21,4 @@ public interface IActivityRepository
 	Task<Activity> ArchiveActivity(Guid activityId, bool isArchived);
 
 	Task<Activity> MarkActivityAsDeleted(Guid activityId, bool isDeleted);
-
-    Task<List<Activity>> GetActivitiesByCategoriesAsync(Guid categoryId, Guid tripId);
 }
