+++ OutTheGC/Interfaces/IActivityService.cs
@@ -21,7 +21,5 @@ public interface IActivityService
     Task<Activity> ArchiveActivity(Guid activityId, bool isArchived);
 
     Task<Activity> MarkActivityAsDeleted(Guid activityId, bool isDeleted);
-
-    Task<List<Activity>> GetActivitiesByCategoriesAsync(Guid categoryId, Guid tripId);
 }
