+++ OutTheGC/Interfaces/ICategoryService.cs
@@ -9,4 +9,6 @@ public interface ICategoryService
     Task<Category> CreateCategoryAsync(Category newCategory);
 
     Task<Category> DeleteCategoryAsync(Guid categoryId);
+
+    Task<List<Activity>> GetActivitiesByCategoriesAsync(Guid categoryId, Guid tripId);
 }
