+++ OutTheGC/Services/CategoryService.cs
@@ -1,5 +1,6 @@
 ﻿using OutTheGC.Interfaces;
 using OutTheGC.Models;
+using OutTheGC.Repositories;
 
 namespace OutTheGC.Services;
 
@@ -26,6 +27,11 @@ public class CategoryService : ICategoryService
     public async Task<Category> DeleteCategoryAsync(Guid categoryId)
 	{
 		return await _categoryRepository.DeleteCategoryAsync(categoryId);
+    }
+
+    public async Task<List<Activity>> GetActivitiesByCategoriesAsync(Guid categoryId, Guid tripId)
+    {
+        return await _categoryRepository.GetActivitiesByCategoriesAsync(categoryId, tripId);
     }
 }
